refactor(web): move applySuggestionCasing to standalone exportable method
