+++ web/src/engine/predictive-text/worker-thread/src/main/model-compositor.ts
@@ -1,12 +1,12 @@
 import * as models from '@keymanapp/models-templates';
-import * as correction from './correction/index.js'
-
 import { KMWString } from '@keymanapp/web-utils';
-
-import TransformUtils from './transformUtils.js';
-import { correctAndEnumerate, dedupeSuggestions, finalizeSuggestions, predictionAutoSelect, processSimilarity, toAnnotatedSuggestion, tupleDisplayOrderSort } from './predict-helpers.js';
-import { detectCurrentCasing, determineModelTokenizer, determineModelWordbreaker, determinePunctuationFromModel } from './model-helpers.js';
 import { LexicalModelTypes } from '@keymanapp/common-types';
+
+import * as correction from './correction/index.js'
+import TransformUtils from './transformUtils.js';
+import { applySuggestionCasing, correctAndEnumerate, dedupeSuggestions, finalizeSuggestions, predictionAutoSelect, processSimilarity, toAnnotatedSuggestion, tupleDisplayOrderSort } from './predict-helpers.js';
+import { detectCurrentCasing, determineModelTokenizer, determineModelWordbreaker, determinePunctuationFromModel } from './model-helpers.js';
+
 import CasingForm = LexicalModelTypes.CasingForm;
 import Context = LexicalModelTypes.Context;
 import Distribution = LexicalModelTypes.Distribution;
@@ -144,7 +144,7 @@ export class ModelCompositor {
     // lexicon for a word.  (Example:  "Apple" the company vs "apple" the fruit.)
     for(let tuple of rawPredictions) {
       if(currentCasing && currentCasing != 'lower') {
-        this.applySuggestionCasing(tuple.prediction.sample, basePrefix, currentCasing);
+        applySuggestionCasing(tuple.prediction.sample, basePrefix, this.lexicalModel, currentCasing);
       }
     }
 
@@ -192,21 +192,6 @@ export class ModelCompositor {
     }
 
     return suggestions;
-  }
-
-  // Responsible for applying casing rules to suggestions.
-  private applySuggestionCasing(suggestion: Suggestion, baseWord: string, casingForm: CasingForm) {
-    // Step 1:  does the suggestion replace the whole word?  If not, we should extend the suggestion to do so.
-    let unchangedLength  = KMWString.length(baseWord) - suggestion.transform.deleteLeft;
-
-    if(unchangedLength > 0) {
-      suggestion.transform.deleteLeft += unchangedLength;
-      suggestion.transform.insert = KMWString.substr(baseWord, 0, unchangedLength) + suggestion.transform.insert;
-    }
-
-    // Step 2: Now that the transform affects the whole word, we may safely apply casing rules.
-    suggestion.transform.insert = this.lexicalModel.applyCasing(casingForm, suggestion.transform.insert);
-    suggestion.displayAs = this.lexicalModel.applyCasing(casingForm, suggestion.displayAs);
   }
 
   acceptSuggestion(suggestion: Suggestion, context: Context, postTransform?: Transform): Reversion {
