+++ web/src/engine/predictive-text/worker-thread/src/tests/mocha/cases/worker-model-compositor.js
@@ -254,154 +254,6 @@ describe('ModelCompositor', function() {
       });
     });
 
-    describe('applySuggestionCasing', function() {
-      let plainApplyCasing = function(caseToApply, text) {
-        switch(caseToApply) {
-          case 'lower':
-            return text.toLowerCase();
-          case 'upper':
-            return text.toUpperCase();
-          case 'initial':
-            return plainApplyCasing('upper', text.charAt(0)) . concat(text.substring(1));
-          default:
-            return text;
-        }
-      };
-
-      var plainCasedModel = new TrieModel(
-        jsonFixture('models/tries/english-1000'), {
-          languageUsesCasing: true,
-          applyCasing: plainApplyCasing,
-          wordBreaker: wordBreakers.default,
-          searchTermToKey: function(text) {
-            // We're dealing with very simple English text; no need to normalize or remove diacritics here.
-            return applyCasing('lower', text);
-          }
-        }
-      );
-
-      it('properly cases suggestions with no suggestion root', function() {
-        var compositor = new ModelCompositor(plainCasedModel, true);
-
-        let suggestion = {
-          transform: {
-            insert: 'the',
-            deleteLeft: 0
-          },
-          displayAs: 'the'
-        };
-
-        compositor.applySuggestionCasing(suggestion, '', 'initial');
-        assert.equal(suggestion.displayAs, 'The');
-        assert.equal(suggestion.transform.insert, 'The');
-
-        suggestion = {
-          transform: {
-            insert: 'thE',
-            deleteLeft: 0
-          },
-          displayAs: 'thE'
-        };
-
-        compositor.applySuggestionCasing(suggestion, '', 'initial');
-        assert.equal(suggestion.displayAs, 'ThE');
-        assert.equal(suggestion.transform.insert, 'ThE');
-
-        suggestion = {
-          transform: {
-            insert: 'the',
-            deleteLeft: 0
-          },
-          displayAs: 'the'
-        };
-
-        compositor.applySuggestionCasing(suggestion, '', 'upper');
-        assert.equal(suggestion.displayAs, 'THE');
-        assert.equal(suggestion.transform.insert, 'THE');
-      });
-
-      it('properly cases suggestions that fully replace the suggestion root', function() {
-        var compositor = new ModelCompositor(plainCasedModel, true);
-
-        let suggestion = {
-          transform: {
-            insert: 'therefore',
-            deleteLeft: 3
-          },
-          displayAs: 'therefore'
-        };
-
-        compositor.applySuggestionCasing(suggestion, 'the', 'initial');
-        assert.equal(suggestion.displayAs, 'Therefore');
-        assert.equal(suggestion.transform.insert, 'Therefore');
-
-        suggestion = {
-          transform: {
-            insert: 'thereFore',
-            deleteLeft: 3
-          },
-          displayAs: 'thereFore'
-        };
-
-        compositor.applySuggestionCasing(suggestion, 'the', 'initial');
-        assert.equal(suggestion.displayAs, 'ThereFore');
-        assert.equal(suggestion.transform.insert, 'ThereFore');
-
-        suggestion = {
-          transform: {
-            insert: 'therefore',
-            deleteLeft: 3
-          },
-          displayAs: 'therefore'
-        };
-
-        compositor.applySuggestionCasing(suggestion, 'the', 'upper');
-        assert.equal(suggestion.displayAs, 'THEREFORE');
-        assert.equal(suggestion.transform.insert, 'THEREFORE');
-      });
-
-      it('properly cases suggestions that do not fully replace the suggestion root', function() {
-        var compositor = new ModelCompositor(plainCasedModel, true);
-
-        let suggestion = {
-          transform: {
-            insert: 'erefore',
-            deleteLeft: 1
-          },
-          displayAs: 'therefore'
-        };
-
-        // When integrated, the 'the' string comes from a wordbreak operation on the current context.
-        compositor.applySuggestionCasing(suggestion, 'the', 'initial');
-        assert.equal(suggestion.displayAs, 'Therefore');
-        assert.equal(suggestion.transform.insert, 'Therefore');
-
-        suggestion = {
-          transform: {
-            insert: 'ereFore',
-            deleteLeft: 1
-          },
-          displayAs: 'thereFore'
-        };
-
-        compositor.applySuggestionCasing(suggestion, 'the', 'initial');
-        assert.equal(suggestion.displayAs, 'ThereFore');
-        assert.equal(suggestion.transform.insert, 'ThereFore');
-
-        suggestion = {
-          transform: {
-            insert: 'erefore',
-            deleteLeft: 1
-          },
-          displayAs: 'therefore'
-        };
-
-        compositor.applySuggestionCasing(suggestion, 'the', 'upper');
-        assert.equal(suggestion.displayAs, 'THEREFORE');
-        assert.equal(suggestion.transform.insert, 'THEREFORE');
-      });
-    });
-
     describe('Model uses default-style keying, no casing', function () {
       var uncasedModel = new TrieModel(
         jsonFixture('models/tries/english-1000'), {
